requirements.txt
fastapi
aiologger[aiofiles]
pip-chill
uvicorn
gunicorn
requests
openai
python-dotenv
pytest
lxml
crewai
crewai-tools